cart finished

+++ js/cart.js
@@ -45,14 +45,19 @@ document.addEventListener('DOMContentLoaded', async()=> {
   const cartItemList = document.querySelector('.cartItemList');
   try {
     cartItemList.innerHTML = ""; // Clear previous items
-    if(cartItems.length != 0){
-      cartItemList.innerHTML = `
+
+    cartItemList.innerHTML = `
+    <div class="paymentContainer" style="display: none">
       <div class="totalPriceContainer">
         <div class="totalPriceText">Total Price:</div>
-        <div class="totalCartPrice">${(cart.totalCartPrice).toFixed(2)} €</div>
+        <div class="totalCartPrice"></div>
       </div>
-    `;
-    }
+      <div paymentBtnContainer>
+        <button class="paymentBtn">Proceed to Payment</button>
+      </div>
+    </div>
+  `;
+
     for(const item of cartItems){
       const itemDiv = document.createElement('div');
       itemDiv.classList.add('itemDiv');
@@ -60,6 +65,10 @@ document.addEventListener('DOMContentLoaded', async()=> {
       try {
         const product = await fetchData(`products/${item.productId}`);
           itemDiv.innerHTML = `
+          <div class="checkboxContainer">
+            <input type="checkbox" class="checkbox" data-item-id="${item.id}"
+            data-item-price="${item.totalPrice}">
+          </div>
           <div class="imageContainer">
             <img class="product-image" src="${product.image}">
           </div>
@@ -87,12 +96,16 @@ document.addEventListener('DOMContentLoaded', async()=> {
   }
 })
 
+//---------------------------------------------------
+// delete cart item 
 document.addEventListener('click',(event)=>{
+  if (event.target.classList.contains("deleteCartItemBtn")){
     const itemId = event.target.dataset.itemId;
     const endPoint = `cartItems/${itemId}`;
     const token = localStorage.getItem("token");
     deleteCartItem(endPoint,token);
     window.location.reload();
+  }
 })
 
 async function deleteCartItem(endPoint,token) {
@@ -104,3 +117,27 @@ async function deleteCartItem(endPoint,token) {
     console.error("Error deleting cartItem:", error);
   }
 }
+
+//---------------------------------------------------
+// payment for selected cart item 
+document.addEventListener('change',(e)=>{
+
+  // Exit if cart is empty
+  if(!cartItems.length) return; 
+
+  const paymentContainer = document.querySelector('.paymentContainer');
+  const checkboxes = document.querySelectorAll('.checkbox');
+  const totalCartPriceElement = document.querySelector(".totalCartPrice");
+
+  let totalPaymentPrice = 0;
+
+  checkboxes.forEach(checkbox =>{
+    if(checkbox.checked){
+      totalPaymentPrice += parseFloat(checkbox.dataset.itemPrice);
+    }
+  })
+  // Show or hide the paymentContainer based on selection
+  paymentContainer.style.display = totalPaymentPrice>0 ? 'block' : 'none';
+  
+  totalCartPriceElement.textContent = `${totalPaymentPrice.toFixed(2)} €`;
+})
